requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
aiohttp==3.10.6
python-multipart==0.0.10
pydantic-settings==2.5.2
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
python-magic==0.4.27
PyPDF2==3.0.1
python-pptx==1.0.2
pdfminer
pdfminer.six
pikepdf
PyMuPDF
fontTools
reportlab
openai
anthropic
transformers
torch